backend/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-jwt-extended>=4.6
flask-sqlalchemy>=3.1
werkzeug>=3.0
torch>=2.0
torchvision>=0.15
pillow>=10.0
matplotlib>=3.7
numpy>=1.24
